newg/pyproject.toml
[tool.poetry]
name = "SamFit - Workout software"
version = "0.1.0"
description = "Be the better of yourself"
authors = ["Samuel L. <chmouel@chmouel.com>"]
license = "Apache-2.0"

[tool.poetry.dependencies]
python = "^3.10"
click = "^8.1.3"
requests = "^2.28.1"
html2text = "^2020.1.16"
python-dateutil = "^2.8.2"

[tool.poetry.group.dev.dependencies]
ipython = "^8.5.0"
mypy = "^0.982"
pylint = "^2.15.3"
ipdb = "^0.13.9"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
